More component refactors
