+++ frontend/components/button/button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react"
 
-import Button from "./Button"
+import Button from "./button"
 
 const meta = {
 	title: "Elements/Button",
