+++ frontend/components/button/index.ts
@@ -1,0 +1,2 @@
+import Button from "./button"
+export default Button
