+++ frontend/components/input/index.ts
@@ -1,0 +1,2 @@
+import Input from "./input"
+export default Input
