+++ frontend/components/input/input.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react"
 
-import Input from "./Input"
+import Input from "./input"
 
 const meta = {
 	title: "Elements/Input",
