+++ frontend/components/member/index.ts
@@ -1,2 +1,2 @@
-import Member from "./Member"
+import Member from "./member"
 export default Member
