+++ frontend/components/posts/index.ts
@@ -1,0 +1,2 @@
+import Posts from "./Posts"
+export default Posts
